feat add send response email to user, move email send functions to util.py and added user feedback messages
