+++ connectionrequest/templates/connectionrequest/response-email/response-email-subject.txt
@@ -1,0 +1,1 @@
+You have recieved a connection response via ReachOut
